# [IOTDB-4006] Improve the performance of Raw Query Without ValueFilter for nonAligned