--- a/server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/process/join/merge/AscTimeComparator.java
+++ b/server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/process/join/merge/AscTimeComparator.java
@@ -16,18 +16,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.db.mpp.execution.operator.process.merge;
+package org.apache.iotdb.db.mpp.execution.operator.process.join.merge;
 
-public class DescTimeComparator implements TimeComparator {
+public class AscTimeComparator implements TimeComparator {
 
-  /** @return if order by time desc, return true if time >= endTime, otherwise false */
+  /** @return if order by time asc, return true if time <= endTime, otherwise false */
   @Override
   public boolean satisfyCurEndTime(long time, long endTime) {
-    return time >= endTime;
+    return time <= endTime;
   }
 
   @Override
   public long getCurrentEndTime(long time1, long time2) {
-    return Math.max(time1, time2);
+    return Math.min(time1, time2);
   }
 }