More details analysis can be found in this [feishu doc](https://apache-iotdb.feishu.cn/docx/doxcnJDbeeC7Tt1aeVHi4YseiNS).

The following is simple test result.

Test Configuration:
1 device, 30sensors(non aligned), each sensor contains 1,000,000 points

Test Machine:
Apple M1 Pro 

Test sql:
We do `selelect * from root.**` query, time cost for old standalone and new standalone:

Old-StandAlone-8-threads: 1552ms
OldStandAlone-one-subThread: 2237ms

ColumnBased-TimeJoin: 3278ms
_**RowBased-TimeJoin:1813ms**_